Clean up rounding up.
Have roundup2() cast the power-of-2 argument to u_int; that way, you
don't have to explicitly define it as an unsigned value in order to
avoid compiler or UBSan complaints about signed integers.

Use it instead of rolling our own rounding-to-a-power-of-2.

+++ netdissect.h
@@ -109,7 +109,7 @@ typedef unsigned char nd_byte;
  * Round up x to a multiple of y; y must be a power of 2.
  */
 #ifndef roundup2
-#define	roundup2(x, y)	(((x)+((y)-1))&(~((y)-1)))
+#define	roundup2(x, y)	(((x)+((u_int)((y)-1)))&(~((u_int)((y)-1))))
 #endif
 
 #include <stdarg.h>
